% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{attr_l}
\alias{attr_l}
\title{Attributes lens}
\usage{
attr_l(name)
}
\arguments{
\item{name}{Name of the attribute to lens into}
}
\value{
A lens that selects the specified attribute
}
\description{
Lens into a named attribute of an object.
}
\examples{
x <- 1:10
attr(x, "label") <- "my_label"
l <- attr_l("label")
view(x, l)
set(x, l, "new_label")
}
