% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{c_l}
\alias{c_l}
\title{Convenience lens for reaching into nested objects}
\usage{
c_l(...)
}
\description{
Convenience function that mirrors \code{\link[purrr:pluck]{purrr::pluck()}}.
}
\seealso{
\code{\link[purrr:pluck]{purrr::pluck()}}
}
