% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-lenses.R
\name{filter_il}
\alias{filter_il}
\title{Filter ilens}
\usage{
filter_il(...)
}
\arguments{
\item{...}{Conditions to filter by}
}
\value{
A lens that filters the specified rows
}
\description{
This function returns an illegal lens that filters according to the specified conditions.
}
\details{
Conditions are evaluated in the context of the data frame.
}
