% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{\%.\%}
\alias{\%.\%}
\title{Compose two lenses}
\usage{
l \%.\% m
}
\arguments{
\item{l}{First lens}

\item{m}{Second lens}
}
\value{
A new lens
}
\description{
The resulting lens first applies the \emph{left} lens, then the right lens.
}
\examples{
d <- list(list(a = 1, b = 2), list(a = 4, b = 9))
l <- index_l(1)
m <- index_l("b")
view(d, l \%.\% m)
}
