% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{id_l}
\alias{id_l}
\title{Identity lens}
\format{
An object of class \code{tinylens::lens} (inherits from \code{S7_object}) of length 1.
}
\usage{
id_l
}
\description{
Trivial identity lens: returns and sets the object itself.
}
\examples{
x <- 1:10
view(x, id_l)
}
\keyword{datasets}
