% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{index_l}
\alias{index_l}
\title{Index lens}
\usage{
index_l(i)
}
\arguments{
\item{i}{Index of the element to lens into}
}
\value{
A lens that selects the specified element
}
\description{
Lens into a single element of a list.
}
\details{
This lens performs indexing using double bracket notation, i.e., \code{x[[i]]}.
}
\examples{
x <- list(a = 1, b = 2)
l <- index_l("a")
view(x, l)
}
