% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lens.R
\name{lens}
\alias{lens}
\title{Create a lens}
\usage{
lens(view = class_missing, set = class_missing)
}
\arguments{
\item{view}{A function that takes an object and returns a value}

\item{set}{A function that takes an object and a value and returns a new object}
}
\description{
A lens is a pair of functions that can be used to view and set a value in an object.
}
