% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{map_l}
\alias{map_l}
\title{Lens into a list or vector}
\usage{
map_l(l, .ptype = NULL)
}
\arguments{
\item{l}{A lens that selects the elements to lens into}

\item{.ptype}{The prototype of the data structure to return}
}
\value{
A lens that selects the specified elements
}
\description{
This lens allows you to access and modify elements of a list or vector
based on their position or a logical condition.
}
\examples{
d <- list(list(a = 1, b = 2), list(a = 4, b = 9))
l <- index_l("a")
view(d, map_l(l))
over_map(d, map_l(l), sqrt)
}
