% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{names_l}
\alias{names_l}
\title{Names lens}
\format{
An object of class \code{tinylens::lens} (inherits from \code{S7_object}) of length 1.
}
\usage{
names_l
}
\description{
Lens into the \code{names} attribute of an object. This uses \code{rlang::names2} to
better handle \code{NULL} names.
}
\examples{
x <- letters[1:10]
names(x) <- letters[1:10]
view(x, names_l)
over(x, names_l, toupper)
}
\keyword{datasets}
