% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{over}
\alias{over}
\title{Modify the focused part of a data structure}
\usage{
over(d, l, f)
}
\arguments{
\item{d}{The data structure to view}

\item{l}{The lens to apply}

\item{f}{The function to apply}
}
\value{
The modified data structure
}
\description{
Modify the focused part of a data structure
}
