% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{over_map}
\alias{over_map}
\title{Map a function over a list lens}
\usage{
over_map(d, l, f)
}
\description{
Apply a function to each element of a list returned by a lens. Using \code{over}
in such cases would require a "lifted" function, which is often unergonomic.
}
