% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-lenses.R
\name{rows_l}
\alias{rows_l}
\title{Rows lens}
\usage{
rows_l(idx)
}
\arguments{
\item{idx}{The rows to select}
}
\value{
A lens that selects the specified rows
}
\description{
This function returns a lens that selects the specified rows.
}
\examples{
d <- data.frame(x = 1:10, y = 11:20, z = 21:30)
l <- rows_l(1:2)
# get the first two rows
view(d, l)
# set the first two rows
set(d, l, 1:2)
}
