% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-lenses.R
\name{select_l}
\alias{select_l}
\title{include verbs.R
include lens.R
Select lens}
\usage{
select_l(...)
}
\arguments{
\item{...}{Columns to select}
}
\value{
A lens that selects the specified columns
}
\description{
This function returns a lens that selects the specified columns.
}
