% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{set}
\alias{set}
\title{Set data through a lens}
\usage{
set(d, l, x)
}
\arguments{
\item{d}{The data structure to set}

\item{l}{The lens to apply}

\item{x}{The value to set}
}
\value{
The modified data structure
}
\description{
This function applies a lens to a data structure and sets the focused part.
}
