% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{slice_l}
\alias{slice_l}
\title{Slice lens}
\usage{
slice_l(idx)
}
\arguments{
\item{idx}{Indices of the elements to lens into}
}
\value{
A lens that selects the specified slice
}
\description{
Lens into a slice of a vector.
}
\details{
This lens performs indexing using single bracket notation, i.e., \code{x[idx]}.
}
\examples{
x <- letters[1:10]
l <- slice_l(1:5)
view(x, l)
}
