% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{c_l}
\alias{c_l}
\title{Lens for accessing and modifying nested elements of a list or vector}
\usage{
c_l(...)
}
\arguments{
\item{...}{A sequence of lenses and/or integers/logical vectors}
}
\value{
A lens that combines all specified lenses (left to right).
}
\description{
Convenience function that mirrors \code{\link[purrr:pluck]{purrr::pluck()}}.
}
\examples{
d <- list(a = list(b = 1, c = 2), b = list(b = 3, c = 4))
l <- c_l("a", "b")
view(d, l)
}
