% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{over}
\alias{over}
\title{Modify data through a lens}
\usage{
over(d, l, f, ...)
}
\arguments{
\item{d}{The data structure to modify}

\item{l}{The lens to apply}

\item{f}{The function to apply}
}
\value{
The modified data structure
}
\description{
This function applies a lens to a data structure and modifies the focused part.
}
