% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframe-lenses.R
\name{rows_l}
\alias{rows_l}
\title{Rows lens}
\usage{
rows_l(idx)
}
\arguments{
\item{idx}{The rows to select}
}
\value{
A lens that selects the specified rows
}
\description{
This function returns a lens that selects the specified rows.
}
