% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{set}
\alias{set}
\title{Set the focused part of a data structure}
\usage{
set(d, l, x)
}
\arguments{
\item{d}{The data structure to view}

\item{l}{The lens to apply}

\item{x}{The value to set}
}
\value{
The modified data structure
}
\description{
Set the focused part of a data structure
}
