% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylens-package.R
\docType{package}
\name{tinylens-package}
\alias{tinylens}
\alias{tinylens-package}
\title{tinylens: Minimal implementation of functional lenses}
\description{
Minimal implementation of functional lenses, inspired by the `lenses` package.
}
\author{
\strong{Maintainer}: Albert Wang \email{albert_z_wang@harvard.edu}

}
\keyword{internal}
