% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinylens-package.R
\docType{package}
\name{tinylens-package}
\alias{tinylens}
\alias{tinylens-package}
\title{tinylens: Minimal Implementation of Functional Lenses}
\description{
Provides utilities to create and use lenses to simplify data manipulation. Lenses are composable getter/setter pairs that provide a functional approach to manipulating deeply nested data structures, e.g., elements within list columns in data frames. The implementation is based on the earlier 'lenses' R package \url{https://github.com/cfhammill/lenses}, which was inspired by the Haskell 'lens' package by Kmett (2012) \url{https://github.com/ekmett/lens}, one of the most widely referenced implementations of lenses. For additional background and history on the theory of lenses, see the 'lens' package wiki: \url{https://github.com/ekmett/lens/wiki/History-of-Lenses}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arbelt/tinylens}
  \item Report bugs at \url{https://github.com/arbelt/tinylens/issues}
}

}
\author{
\strong{Maintainer}: Albert Wang \email{albert_z_wang@harvard.edu} [copyright holder]

}
\keyword{internal}
