% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{vec_data_l}
\alias{vec_data_l}
\title{Lens into vector data}
\format{
An object of class \code{tinylens::lens} (inherits from \code{S7_object}) of length 1.
}
\usage{
vec_data_l
}
\description{
Allows mutation of vector data while preserving attributes, e.g., labels.
}
\keyword{datasets}
