% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\docType{data}
\name{vec_data_l}
\alias{vec_data_l}
\title{Vector data lens}
\format{
An object of class \code{tinylens::lens} (inherits from \code{S7_object}) of length 1.
}
\usage{
vec_data_l
}
\description{
Allows mutation of vector data while preserving attributes, e.g., labels or names.
}
\examples{
x <- letters[1:10]
names(x) <- letters[1:10]
# toy function that strips names; most functions from `stringr` do this
f <- function(x) toupper(unname(x))
# apply the function without losing attributes
over(x, vec_data_l, f)
}
\keyword{datasets}
