% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbs.R
\name{view}
\alias{view}
\title{View data through a lens}
\usage{
view(d, l, ...)
}
\arguments{
\item{d}{The data structure to view}

\item{l}{The lens to apply}
}
\value{
The part of the data structure focused by the lens
}
\description{
This function applies a lens to a data structure and returns the focused part.
}
