% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base-lenses.R
\name{where_il}
\alias{where_il}
\title{Predicate ilens}
\usage{
where_il(p)
}
\arguments{
\item{p}{A predicate function}
}
\value{
A lens that selects the elements that satisfy the predicate
}
\description{
Illegal lens into elements of a vector that satisfy a predicate.
}
